Restructuring
Using a class
